Sort main.py into cogs and helpers
