+++ helpers/CCCommand.py
@@ -1,0 +1,13 @@
+from sqlalchemy import Column, String, DeclarativeMeta
+
+class CCCommand(DeclarativeMeta):
+    """
+    Makes object that database undserstands
+    """
+    __tablename__ = "imageCommands"
+
+    #Has a column for the ID, name, and responce
+    #id = Column(Integer, primary_key=True)
+    name = Column(String, primary_key=True)
+    responce = Column(String)
+    category = Column(String)
