+++ helpers/icon_animator.py
